expose admin routes and remove unwraps
