+++ server/src/routes/admin.rs
@@ -17,6 +17,10 @@ pub fn router() -> Router<AppState> {
         .route("/all_users", get(user_admin::get_all_users))
         .route("/delete_user", delete(user_admin::delete_user));
 
-    let admin_route = Router::new().nest("/user_admin", user_admin);
+    let group_admin = Router::new()
+        .route("/history", get(group_admin::get_groups_metrics))
+        .route("/transfer_metrics", get(group_admin::get_payments_totals))
+    ;
+    let admin_route = Router::new().nest("/user_admin", user_admin).nest("/group_admin", group_admin);
     admin_route
 }
